backend/api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import RestaurantViewSet, CartViewSet, OrderViewSet, LoginView, LogoutView

router = DefaultRouter()
router.register(r'restaurants', RestaurantViewSet, basename='restaurant')
router.register(r'cart', CartViewSet, basename='cart')
router.register(r'orders', OrderViewSet, basename='order')

urlpatterns = [
    path('login/', LoginView.as_view(), name='login'),
    path('logout/', LogoutView.as_view(), name='logout'),
    path('', include(router.urls)),
]