backend/pyproject.toml
[tool.poetry]
name = "delivery"
version = "0.1.0"
description = ""
authors = ["Xei201 <xei201@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.2"
redis = "^5.0.7"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^70.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
